requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
chromadb>=0.4.22
pypdf>=4.0.0